(fix) Fixed any user being able to delete any saved seacrh

+++ src/modules/search/resolvers/remove-saved-search.ts
@@ -3,7 +3,14 @@ import House from "../../../models/house";
 import SavedSearch from "../../../models/saved-search";
 
 const removeSavedHouse = async (_, args, { user }: { user: User }) => {
-  return await SavedSearch.query().deleteById(args.id);
+  //fetch the saved search
+  const savedSearch = await SavedSearch.query().findOne("id", args.id);
+  //check if current user saved it
+  if (savedSearch && savedSearch.userId == user.id) {
+    //delete saved search
+    return await SavedSearch.query().deleteById(args.id);
+  }
+  throw Error("Saved Search not found!");
 };
 
 export default removeSavedHouse;
